Add user ID filter to database queries to isolate data.

+++ backend/main.py
@@ -102,8 +102,8 @@ app.add_middleware(
 @app.get("/transactions")
 def get_transactions(page: int = 0, pageSize: int = 10, user = Depends(get_firebase_user)):
     database = Session()
-    database_results = database.query(Transaction).offset(page * pageSize).limit(pageSize).all()
-    count = database.query(func.count(Transaction.id)).scalar()
+    database_results = database.query(Transaction).filter_by(id=id, user_id=user["user_id"]).offset(page * pageSize).limit(pageSize).all()
+    count = database.query(func.count(Transaction.id)).filter_by(id=id, user_id=user["user_id"]).scalar()
     database.close()
     return {
         "count": count,
@@ -116,8 +116,8 @@ def get_transactions(page: int = 0, pageSize: int = 10, user = Depends(get_fireb
 @app.get("/reports")
 def get_reports(page: int = 0, pageSize: int = 10, user = Depends(get_firebase_user)):
     database = Session()
-    database_results = database.query(Report).order_by(desc(Report.requested_date)).offset(page * pageSize).limit(pageSize).all()
-    count = database.query(func.count(Report.id)).scalar()
+    database_results = database.query(Report).filter_by(id=id, user_id=user["user_id"]).order_by(desc(Report.requested_date)).offset(page * pageSize).limit(pageSize).all()
+    count = database.query(func.count(Report.id)).filter_by(id=id, user_id=user["user_id"]).scalar()
     database.close()
     return {
         "count": count,
@@ -166,7 +166,7 @@ async def post_transactions(file: UploadFile, user = Depends(get_firebase_user))
 @app.get("/transaction/{id}")
 def delete_transaction(id, user = Depends(get_firebase_user)):
     database = Session()
-    database.query(Transaction).filter_by(id = id).delete()
+    database.query(Transaction).filter_by(id=id, user_id=user["user_id"]).delete()
     database.commit()
     database.close()
